+++ DebugStringTest.java
@@ -1,0 +1,6 @@
+public class DebugStringTest {
+    public void test() {
+        String test = "This has + + inside";
+        i + +;
+    }
+}